\name{DirichletMultinomial-package}
\alias{DirichletMultinomial-package}
\docType{package}
\title{

  Dirichlet-Multinomial Mixture Model Machine Learning for Microbiome
  Data

}
\description{

  Dirichlet-multinomial mixture models can be used to describe
  variability in microbial metagenomic data. This package is an
  interface to code originally made available by Holmes, Harris, and
  Qunice, 2012, PLoS ONE 7(2): 1-15.

}
\details{

  The estimation routine is from the LGPL-licensed (as stated on the
  corresponding googlecode page) source
  \url{http://microbedmm.googlecode.com/files/MicrobeDMMv1.0.tar.gz},
  retrieved 17 Feburary 2012.

  The algorithm is described in Holmes I, Harris K, Quince C, 2012
  Dirichlet Multinomial Mixtures: Generative Models for Microbial
  Metagenomics. PLoS ONE 7(2): e30126. doi:10.1371/journal.pone.0030126.

}
\author{Maintainer: Martin Morgan \url{mailto:mtmorgan@fhcrc.org}}
\keyword{package}
